+++ packages/terra-signature/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Update wdio tests to remove deprecated Terra.it helpers
+
 ## 2.32.0 - (August 4, 2020)
 
 * Changed